remove circular dependency eww
