+++ frontend/src/embedding-sdk-bundle/components/private/SdkInternalNavigation/context.tsx
@@ -1,0 +1,60 @@
+import { createContext, useContext } from "react";
+
+import type { SdkDashboardId } from "embedding-sdk-bundle/types/dashboard";
+import type {
+  NavigateToNewCardParams,
+  SdkQuestionId,
+} from "embedding-sdk-bundle/types/question";
+import type { ParameterValues } from "metabase/embedding-sdk/types/dashboard";
+
+export type SdkInternalNavigationEntry =
+  | {
+      type: "dashboard";
+      id: SdkDashboardId;
+      name: string;
+      parameters?: ParameterValues;
+    }
+  | {
+      type: "question";
+      id: SdkQuestionId;
+      name: string;
+      parameters?: ParameterValues;
+    }
+  | {
+      type: "adhoc-question";
+      /** The URL path for the ad-hoc question (e.g., /question#... with serialized card) */
+      questionPath: string;
+      name: string;
+    };
+
+export type SdkInternalNavigationContextValue = {
+  stack: SdkInternalNavigationEntry[];
+  push: (entry: SdkInternalNavigationEntry) => void;
+  pop: () => void;
+  currentEntry: SdkInternalNavigationEntry | undefined;
+  canGoBack: boolean;
+  previousEntry: SdkInternalNavigationEntry | undefined;
+  navigateToNewCard: (params: NavigateToNewCardParams) => Promise<void>;
+  initWithDashboard: (dashboard: { id: number; name: string }) => void;
+};
+
+export const SdkInternalNavigationContext =
+  createContext<SdkInternalNavigationContextValue | null>(null);
+
+export const useSdkInternalNavigation = () => {
+  const ctx = useContext(SdkInternalNavigationContext);
+  if (!ctx) {
+    throw new Error(
+      "useSdkInternalNavigation must be used within SdkInternalNavigationProvider",
+    );
+  }
+  return ctx;
+};
+
+/**
+ * Optional version of useSdkInternalNavigation that returns null if outside provider.
+ * Useful for components that may be rendered outside the navigation context.
+ */
+export const useSdkInternalNavigationOptional = () => {
+  return useContext(SdkInternalNavigationContext);
+};
