+++ frontend/src/embedding-sdk-bundle/hooks/private/use-sdk-internal-navigation-back.ts
@@ -1,4 +1,4 @@
-import { useSdkInternalNavigation } from "embedding-sdk-bundle/components/private/SdkInternalNavigationProvider";
+import { useSdkInternalNavigation } from "embedding-sdk-bundle/components/private/SdkInternalNavigation/context";
 
 export const useSdkInternalNavigationBack = () => {
   const { previousEntry, canGoBack, pop } = useSdkInternalNavigation();
